Spring boot update (#19)
* Changed spring mvc to spring boot

* Added fix for i18n in spring boot

* Initializer is not required with spring boot

* Updated readme.md with spring-boot info

* Updated backend endpoint with proper use of restful spec

* Updated ui spec for updating and deleting question

* Added update for question type api endpoint

* Fixed answer not updating properly

+++ pdf-app/src/main/java/com/bitMiners/pdf/services/impl/QuestionTypeServiceImpl.java
@@ -1,12 +1,14 @@
 package com.bitMiners.pdf.services.impl;
 
 import com.bitMiners.pdf.domain.QuestionType;
+import com.bitMiners.pdf.exceptions.PdfApiException;
 import com.bitMiners.pdf.repositories.QuestionTypeRepository;
 import com.bitMiners.pdf.services.QuestionTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class QuestionTypeServiceImpl implements QuestionTypeService {
@@ -18,11 +20,19 @@ public class QuestionTypeServiceImpl implements QuestionTypeService {
         return questionTypeRepository.add(questionType);
     }
 
-    public Long updateQuestionType(QuestionType questionType) {
-        return questionTypeRepository.update(questionType) ? 1L : 0;
+    public QuestionType updateQuestionType(QuestionType questionType) {
+        QuestionType dbRecord = this.questionTypeRepository.findOne(questionType.getId());
+        if (Objects.isNull(dbRecord)) {
+            throw new PdfApiException("No Question Type found with id: " + questionType.getId(), 404);
+        }
+        return questionTypeRepository.update(questionType);
     }
 
     public void deleteQuestionType(Long id) {
+        QuestionType dbRecord = this.questionTypeRepository.findOne(id);
+        if (Objects.isNull(dbRecord)) {
+            throw new PdfApiException("No Question Type found with id: " + id, 404);
+        }
         questionTypeRepository.delete(id);
     }
